+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/tax_category/TaxRateImpl.java
@@ -39,6 +39,8 @@ public class TaxRateImpl implements TaxRate, ModelBase {
 
     private java.util.List<com.commercetools.api.models.tax_category.SubRate> subRates;
 
+    private com.commercetools.api.models.tax_category.TaxRoundingTarget taxRoundingTarget;
+
     /**
      * create instance with all properties
      */
@@ -47,7 +49,8 @@ public class TaxRateImpl implements TaxRate, ModelBase {
             @JsonProperty("name") final String name, @JsonProperty("amount") final Double amount,
             @JsonProperty("includedInPrice") final Boolean includedInPrice,
             @JsonProperty("country") final String country, @JsonProperty("state") final String state,
-            @JsonProperty("subRates") final java.util.List<com.commercetools.api.models.tax_category.SubRate> subRates) {
+            @JsonProperty("subRates") final java.util.List<com.commercetools.api.models.tax_category.SubRate> subRates,
+            @JsonProperty("taxRoundingTarget") final com.commercetools.api.models.tax_category.TaxRoundingTarget taxRoundingTarget) {
         this.id = id;
         this.key = key;
         this.name = name;
@@ -56,6 +59,7 @@ public class TaxRateImpl implements TaxRate, ModelBase {
         this.country = country;
         this.state = state;
         this.subRates = subRates;
+        this.taxRoundingTarget = taxRoundingTarget;
     }
 
     /**
@@ -129,6 +133,15 @@ public class TaxRateImpl implements TaxRate, ModelBase {
         return this.subRates;
     }
 
+    /**
+     *  <p>Determines which of the net price and the tax amount the <code>taxRoundingMode</code> of the Cart or Order is applied to, when this TaxRate is included in the price. Ignored if <code>includedInPrice</code> is <code>false</code>.</p>
+     *  <p>Always returned by the API. Can be omitted when a TaxRate is supplied as input, such as in <a href="https://docs.commercetools.com/apis/ctp:api:type:OrderImportDraft" rel="nofollow">OrderImportDraft</a>, and then defaults to <code>Net</code>.</p>
+     */
+
+    public com.commercetools.api.models.tax_category.TaxRoundingTarget getTaxRoundingTarget() {
+        return this.taxRoundingTarget;
+    }
+
     public void setId(final String id) {
         this.id = id;
     }
@@ -165,6 +178,11 @@ public class TaxRateImpl implements TaxRate, ModelBase {
         this.subRates = subRates;
     }
 
+    public void setTaxRoundingTarget(
+            final com.commercetools.api.models.tax_category.TaxRoundingTarget taxRoundingTarget) {
+        this.taxRoundingTarget = taxRoundingTarget;
+    }
+
     @Override
     public boolean equals(Object o) {
         if (this == o)
@@ -183,6 +201,7 @@ public class TaxRateImpl implements TaxRate, ModelBase {
                 .append(country, that.country)
                 .append(state, that.state)
                 .append(subRates, that.subRates)
+                .append(taxRoundingTarget, that.taxRoundingTarget)
                 .append(id, that.id)
                 .append(key, that.key)
                 .append(name, that.name)
@@ -191,6 +210,7 @@ public class TaxRateImpl implements TaxRate, ModelBase {
                 .append(country, that.country)
                 .append(state, that.state)
                 .append(subRates, that.subRates)
+                .append(taxRoundingTarget, that.taxRoundingTarget)
                 .isEquals();
     }
 
@@ -204,6 +224,7 @@ public class TaxRateImpl implements TaxRate, ModelBase {
                 .append(country)
                 .append(state)
                 .append(subRates)
+                .append(taxRoundingTarget)
                 .toHashCode();
     }
 
@@ -217,6 +238,7 @@ public class TaxRateImpl implements TaxRate, ModelBase {
                 .append("country", country)
                 .append("state", state)
                 .append("subRates", subRates)
+                .append("taxRoundingTarget", taxRoundingTarget)
                 .build();
     }