+++ changes.md
@@ -157,6 +157,7 @@
 - added type `StoreSetRefundPolicyUrlAction`
 - added type `StoreSetShippingPolicyUrlAction`
 - added type `StoreSetTermsOfServiceUrlAction`
+- added type `TaxRoundingTarget`
 - added type `TypeRemoveEnumValuesAction`
 - added type `TypeRemoveLocalizedEnumValuesAction`
 - added type `VariantMoveImageToPositionAction`
@@ -171,6 +172,7 @@
 - added property `directDiscountsIgnoreCartDiscounts` to type `CartDraft`
 - added property `participateInBestDealSelection` to type `DirectDiscount`
 - added property `participateInBestDealSelection` to type `DirectDiscountDraft`
+- added property `taxRoundingTarget` to type `ExternalTaxRateDraft`
 - added property `estimatedDelivery` to type `ShippingInfo`
 - added property `estimatedDelivery` to type `CartAddShippingMethodAction`
 - added property `estimatedDelivery` to type `CartSetCustomShippingMethodAction`
@@ -203,6 +205,8 @@
 - added property `carrier` to type `ShippingMethodDraft`
 - added property `storefront` to type `Store`
 - added property `storefront` to type `StoreDraft`
+- added property `taxRoundingTarget` to type `TaxRate`
+- added property `taxRoundingTarget` to type `TaxRateDraft`
 - added property `warnings` to type `Variant`
 - added property `categories` to type `VariantProjection`
 - added property `categoryOrderHints` to type `VariantProjection`